requirements.txt
attrs==21.4.0
black==22.1.0
click==8.0.3
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
fakeredis==1.7.1
Flask==2.0.2
flask-mongoengine==1.0.0
flask-redis==0.4.0
Flask-WTF==1.0.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.23.1
mongomock==4.0.0
mypy-extensions==0.4.3
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pymongo==3.12.3
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
sentinels==1.0.0
six==1.16.0
sortedcontainers==2.4.0
tomli==2.0.1
typeguard==2.13.3
typing-extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
